+++ changelogs/CHANGELOG.md
@@ -1,3 +1,8 @@
+### Cobbled Paths v0.1.3
+
+- Added Nether Brick and Blackstone paths
+- Added cracked variants of most paths and decorated variants of a few
+
 ### Cobbled Paths v0.1.2
 
 Changed the logic used by path blocks to be more logical; they should now only wrap around edges when there are path blocks next to them.